refactor: formatElapsed
