+++ packages/maskbook/src/_locales/en/messages.json
@@ -327,12 +327,12 @@
     "plugin_red_packet_description_empty": "The red packet is empty.",
     "plugin_red_packet_description_broken": "The red packet is broken.",
     "plugin_red_packet_description_failover": "From {{name}} / {{shares}} shares / {{total}} {{symbol}}",
-    "plugin_red_packet_relative_time_days_ago": "{{days}} days ago",
-    "plugin_red_packet_relative_time_hours_ago": "{{hours}} hours ago",
-    "plugin_red_packet_relative_time_minutes_ago": "{{minutes}} minutes ago",
-    "plugin_red_packet_relative_time_months_ago": "{{months}} months ago",
-    "plugin_red_packet_relative_time_seconds_ago": "{{seconds}} seconds ago",
-    "plugin_red_packet_relative_time_years_ago": "{{years}} years ago",
+    "relative_time_days_ago": "{{days}} days ago",
+    "relative_time_hours_ago": "{{hours}} hours ago",
+    "relative_time_minutes_ago": "{{minutes}} minutes ago",
+    "relative_time_months_ago": "{{months}} months ago",
+    "relative_time_seconds_ago": "{{seconds}} seconds ago",
+    "relative_time_years_ago": "{{years}} years ago",
     "plugin_gitcoin_readme": "By using this service, you will also be contributing 5% of your contribution to the <fund>Gitcoin grants development fund</fund>.",
     "plugin_gitcoin_readme_fund_link": "https://gitcoin.co/grants/86/gitcoin-sustainability-fund",
     "plugin_gitcoin_preview_card_fund": "Fund this grant",
@@ -374,6 +374,7 @@
     "plugin_trader_table_pair": "Pair",
     "plugin_trader_table_price": "Price",
     "plugin_trader_table_volume": "Volume (24h)",
+    "plugin_trader_table_updated": "Updated",
     "plugin_trader_finding_price": "Finding price…",
     "plugin_trader_error_amount_absence": "Enter an amount",
     "plugin_trader_error_insufficient_balance": "Insufficient {{symbol}} balance",
