+++ packages/maskbook/src/_locales/ja/messages.json
@@ -327,12 +327,12 @@
     "plugin_red_packet_description_empty": "期限切れです",
     "plugin_red_packet_description_broken": "破損しています",
     "plugin_red_packet_description_failover": "{{amount}} {{symbol}} / {{shares}} 株",
-    "plugin_red_packet_relative_time_days_ago": "{{days}}日前",
-    "plugin_red_packet_relative_time_hours_ago": "{{hours}}時間前",
-    "plugin_red_packet_relative_time_minutes_ago": "{{minutes}}分前",
-    "plugin_red_packet_relative_time_months_ago": "{{months}}月前",
-    "plugin_red_packet_relative_time_seconds_ago": "{{seconds}}秒前",
-    "plugin_red_packet_relative_time_years_ago": "{{years}}年前",
+    "relative_time_days_ago": "{{days}}日前",
+    "relative_time_hours_ago": "{{hours}}時間前",
+    "relative_time_minutes_ago": "{{minutes}}分前",
+    "relative_time_months_ago": "{{months}}月前",
+    "relative_time_seconds_ago": "{{seconds}}秒前",
+    "relative_time_years_ago": "{{years}}年前",
     "plugin_gitcoin_readme": "このサービスを利用するにあたり、寄付金額の 5% が <fund>Gitcoin grants development fund</fund> に寄付されます",
     "plugin_gitcoin_readme_fund_link": "https://gitcoin.co/grants/86/gitcoin-sustainability-fund",
     "plugin_gitcoin_preview_card_fund": "この Grant に寄付する",
@@ -374,6 +374,7 @@
     "plugin_trader_table_pair": "通貨ペア",
     "plugin_trader_table_price": "価格",
     "plugin_trader_table_volume": "取引高（24時間）",
+    "plugin_trader_table_updated": "",
     "plugin_trader_finding_price": "",
     "plugin_trader_error_amount_absence": "",
     "plugin_trader_error_insufficient_balance": "",
