+++ packages/maskbook/src/_locales/zh/messages.json
@@ -327,12 +327,12 @@
     "plugin_red_packet_description_empty": "已全部被別人領取",
     "plugin_red_packet_description_broken": "",
     "plugin_red_packet_description_failover": "{{amount}} {{symbol}} / {{shares}} 份",
-    "plugin_red_packet_relative_time_days_ago": "",
-    "plugin_red_packet_relative_time_hours_ago": "",
-    "plugin_red_packet_relative_time_minutes_ago": "",
-    "plugin_red_packet_relative_time_months_ago": "",
-    "plugin_red_packet_relative_time_seconds_ago": "",
-    "plugin_red_packet_relative_time_years_ago": "",
+    "relative_time_days_ago": "",
+    "relative_time_hours_ago": "",
+    "relative_time_minutes_ago": "",
+    "relative_time_months_ago": "",
+    "relative_time_seconds_ago": "",
+    "relative_time_years_ago": "",
     "plugin_gitcoin_readme": "在使用本服務時，你的捐贈的 5% 將會被用於支持 <fund>Gitcoin grants development fund</fund>",
     "plugin_gitcoin_readme_fund_link": "https://gitcoin.co/grants/86/gitcoin-sustainability-fund",
     "plugin_gitcoin_preview_card_fund": "為此 Grant 捐贈",
@@ -374,6 +374,7 @@
     "plugin_trader_table_pair": "交易對",
     "plugin_trader_table_price": "價格",
     "plugin_trader_table_volume": "交易量 （24小時）",
+    "plugin_trader_table_updated": "",
     "plugin_trader_finding_price": "",
     "plugin_trader_error_amount_absence": "",
     "plugin_trader_error_insufficient_balance": "",
