+++ packages/maskbook/src/plugins/RedPacket/hooks/useClaimCallback.ts
@@ -5,7 +5,7 @@ import { useRedPacketContract } from '../contracts/useRedPacketContract'
 import { useTransactionState, TransactionStateType } from '../../../web3/hooks/useTransactionState'
 import type { Tx } from '../../../contracts/types'
 import { addGasMargin } from '../../../web3/helpers'
-import { RedPacketRPC } from '../helpers'
+import { RedPacketRPC } from '../messages'
 
 export function useClaimCallback(from: string, id?: string, password?: string) {
     const [claimState, setClaimState] = useTransactionState()
