+++ packages/maskbook/src/plugins/RedPacket/services.ts
@@ -2,7 +2,7 @@ import * as jwt from 'jsonwebtoken'
 import { sha3 } from 'web3-utils'
 import type { RedPacketRecord, RedPacketJSONPayload, History } from './types'
 import { RED_PACKET_HISTORY_URL } from './constants'
-import { RedPacketMessage } from './helpers'
+import { RedPacketMessage } from './messages'
 import * as database from './database'
 import { resolveChainName } from '../../web3/pipes'
 import Services from '../../extension/service'
